Add support for Git LFS with the default client. Fixes #8.
Uses the git-lfs installed on the current server if possible, otherwise displays an error message with instructions on how to install Git LFS. No change for non-LFS repositories.

This also changes LibGit2Sharp to use the same copy-from-working-directory method as the native client version uses because the object database contains LFS pointer files, not the large files they point to.